requirements.txt
asgiref==3.2.3
attrs==19.3.0
coverage==5.0.3
Django==3.0.1
django-cors-headers==3.2.0
django-js-asset==1.2.2
django-mptt==0.10.0
djangorestframework==3.11.0
Faker==0.9.1
importlib-metadata==1.4.0
mixer==6.1.3
more-itertools==8.1.0
packaging==20.0
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
pyparsing==2.4.6
pytest==5.3.4
pytest-cov==2.8.1
pytest-django==3.8.0
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.2
wcwidth==0.1.8
zipp==2.0.0
